refactor: SecurityUtil 수정사항 반영
- Security context null 에러 타입 Unauthorized 로 변경
- SecurityUtil package util 로 변경
- getCurrentMemberId -> getCurrentUserId
