+++ post-service/src/main/java/com/ncns/sns/post/service/PostService.java
@@ -1,6 +1,6 @@
 package com.ncns.sns.post.service;
 
-import com.ncns.sns.post.common.SecurityUtil;
+import com.ncns.sns.post.util.SecurityUtil;
 import com.ncns.sns.post.domain.*;
 import com.ncns.sns.post.dto.request.CreatePostRequestDto;
 import com.ncns.sns.post.dto.request.UpdatePostRequestDto;
@@ -83,13 +83,13 @@ public class PostService {
 
     @Transactional
     public String requestLikePost(Long postId) {
-        Long likeData = likeRepository.isLiked(postId, SecurityUtil.getCurrentMemberId());
+        Long likeData = likeRepository.isLiked(postId, SecurityUtil.getCurrentUserId());
         return likeData == null ? like(postId) : disLike(likeData, postId);
     }
 
     private Post checkAuthorization(Long postId) {
         Post post = getPostById(postId);
-        if (!post.getUserId().equals(SecurityUtil.getCurrentMemberId())) {
+        if (!post.getUserId().equals(SecurityUtil.getCurrentUserId())) {
             throw new BadRequestException(ResponseType.POST_NOT_AUTHOR);
         }
         return post;
@@ -109,7 +109,7 @@ public class PostService {
     private String like(Long postId) {
         Like like = Like.builder()
                 .postId(postId)
-                .userId(SecurityUtil.getCurrentMemberId())
+                .userId(SecurityUtil.getCurrentUserId())
                 .build();
         likeRepository.save(like);
         PostCount postCount = postCountRepository.findByPostId(postId);
