+++ post-service/src/main/java/com/ncns/sns/post/util/SecurityUtil.java
@@ -1,5 +1,7 @@
-package com.ncns.sns.post.common;
+package com.ncns.sns.post.util;
 
+import dev.ncns.sns.common.domain.ResponseType;
+import dev.ncns.sns.common.exception.UnauthorizedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 
@@ -7,10 +9,10 @@ public class SecurityUtil {
     private SecurityUtil() {
     }
 
-    public static Long getCurrentMemberId() {
+    public static Long getCurrentUserId() {
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         if (authentication == null || authentication.getName() == null) {
-            throw new RuntimeException("Security Context 에 인증 정보가 없습니다.");
+            throw new UnauthorizedException(ResponseType.REQUEST_UNAUTHORIZED);
 
         }
         return Long.parseLong(authentication.getName());
